Moved test to seperate  test assembly
